doppalf-ai/requirements.txt
cohere==5.5.0
fastapi==0.110.0
llama-index==0.10.34
llama-index-embeddings-cohere==0.1.8
llama-index-llms-cohere==0.2.0
llama-index-postprocessor-cohere-rerank==0.1.6
llama-index-vector-stores-qdrant==0.2.8
python-dotenv==1.0.1
qdrant-client==1.9.1
uvicorn==0.29.0
